requirements.txt
celery==5.5.3
Django==4.2.24
django-cors-headers==4.4.0
django-environ==0.11.2
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
flake8==7.1.2
gunicorn==23.0.0
psycopg2-binary==2.9.10
pyflakes==3.2.0
PyJWT==2.9.0
pytest==8.3.5
pytest-django==4.11.1
python-decouple==3.8
redis==6.1.1
six==1.17.0
